+++ sickrage/libs/rtorrentlib/torrent.py
@@ -18,18 +18,18 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import rtorrent.compat
-import rtorrent.file
+import rtorrentlib.compat
+import rtorrentlib.file
 # from rtorrent.rpc import Method
-import rtorrent.peer
-import rtorrent.rpc
-import rtorrent.tracker
-from rtorrent.common import safe_repr
+import rtorrentlib.peer
+import rtorrentlib.rpc
+import rtorrentlib.tracker
+from rtorrentlib.common import safe_repr
 
-Peer = rtorrent.peer.Peer
-Tracker = rtorrent.tracker.Tracker
-File = rtorrent.file.File
-Method = rtorrent.rpc.Method
+Peer = rtorrentlib.peer.Peer
+Tracker = rtorrentlib.tracker.Tracker
+File = rtorrentlib.file.File
+Method = rtorrentlib.rpc.Method
 
 
 class Torrent:
@@ -67,10 +67,10 @@ class Torrent:
         @note: also assigns return value to self.peers
         """
         self.peers = []
-        retriever_methods = [m for m in rtorrent.peer.methods
+        retriever_methods = [m for m in rtorrentlib.peer.methods
                              if m.is_retriever() and m.is_available(self._rt_obj)]
         # need to leave 2nd arg empty (dunno why)
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         m.add("p.multicall", self.info_hash, "",
               *[method.rpc_call + "=" for method in retriever_methods])
 
@@ -80,7 +80,7 @@ class Torrent:
             results_dict = {}
             # build results_dict
             for m, r in zip(retriever_methods, result):
-                results_dict[m.varname] = rtorrent.rpc.process_result(m, r)
+                results_dict[m.varname] = rtorrentlib.rpc.process_result(m, r)
 
             self.peers.append(Peer(
                 self._rt_obj, self.info_hash, **results_dict))
@@ -96,11 +96,11 @@ class Torrent:
         @note: also assigns return value to self.trackers
         """
         self.trackers = []
-        retriever_methods = [m for m in rtorrent.tracker.methods
+        retriever_methods = [m for m in rtorrentlib.tracker.methods
                              if m.is_retriever() and m.is_available(self._rt_obj)]
 
         # need to leave 2nd arg empty (dunno why)
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         m.add("t.multicall", self.info_hash, "",
               *[method.rpc_call + "=" for method in retriever_methods])
 
@@ -110,7 +110,7 @@ class Torrent:
             results_dict = {}
             # build results_dict
             for m, r in zip(retriever_methods, result):
-                results_dict[m.varname] = rtorrent.rpc.process_result(m, r)
+                results_dict[m.varname] = rtorrentlib.rpc.process_result(m, r)
 
             self.trackers.append(Tracker(
                 self._rt_obj, self.info_hash, **results_dict))
@@ -127,18 +127,18 @@ class Torrent:
         """
 
         self.files = []
-        retriever_methods = [m for m in rtorrent.file.methods
+        retriever_methods = [m for m in rtorrentlib.file.methods
                              if m.is_retriever() and m.is_available(self._rt_obj)]
         # 2nd arg can be anything, but it'll return all files in torrent
         # regardless
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         m.add("f.multicall", self.info_hash, "",
               *[method.rpc_call + "=" for method in retriever_methods])
 
         results = m.call()[0]  # only sent one call, only need first result
 
         offset_method_index = retriever_methods.index(
-            rtorrent.rpc.find_method("f.get_offset"))
+            rtorrentlib.rpc.find_method("f.get_offset"))
 
         # make a list of the offsets of all the files, sort appropriately
         offset_list = sorted([r[offset_method_index] for r in results])
@@ -147,7 +147,7 @@ class Torrent:
             results_dict = {}
             # build results_dict
             for m, r in zip(retriever_methods, result):
-                results_dict[m.varname] = rtorrent.rpc.process_result(m, r)
+                results_dict[m.varname] = rtorrentlib.rpc.process_result(m, r)
 
             # get proper index positions for each file (based on the file
             # offset)
@@ -165,7 +165,7 @@ class Torrent:
         Also doesn't restart after directory is set, that must be called
         separately.
         """
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.try_stop")
         self.multicall_add(m, "d.set_directory", d)
 
@@ -178,13 +178,13 @@ class Torrent:
         Also doesn't restart after directory is set, that must be called
         separately.
         """
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.try_stop")
         self.multicall_add(m, "d.set_directory_base", d)
 
     def start(self):
         """Start the torrent"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.try_start")
         self.multicall_add(m, "d.is_active")
 
@@ -193,7 +193,7 @@ class Torrent:
 
     def stop(self):
         """"Stop the torrent"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.try_stop")
         self.multicall_add(m, "d.is_active")
 
@@ -202,21 +202,21 @@ class Torrent:
 
     def pause(self):
         """Pause the torrent"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.pause")
 
         return(m.call()[-1])
 
     def resume(self):
         """Resume the torrent"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.resume")
 
         return(m.call()[-1])
 
     def close(self):
         """Close the torrent and it's files"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.close")
 
         return(m.call()[-1])
@@ -225,14 +225,14 @@ class Torrent:
         """Delete the torrent
 
         @note: doesn't delete the downloaded files"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.erase")
 
         return(m.call()[-1])
 
     def check_hash(self):
         """(Re)hash check the torrent"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.check_hash")
 
         return(m.call()[-1])
@@ -250,7 +250,7 @@ class Torrent:
 
         @return: None
         """
-        multicall = rtorrent.rpc.Multicall(self)
+        multicall = rtorrentlib.rpc.Multicall(self)
         retriever_methods = [m for m in methods
                              if m.is_retriever() and m.is_available(self._rt_obj)]
         for method in retriever_methods:
@@ -272,14 +272,14 @@ class Torrent:
         else:
             call = "d.accepting_seeders.disable"
 
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, call)
 
         return(m.call()[-1])
 
     def announce(self):
         """Announce torrent info to tracker(s)"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.tracker_announce")
 
         return(m.call()[-1])
@@ -300,7 +300,7 @@ class Torrent:
         """
 
         self._assert_custom_key_valid(key)
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
 
         field = "custom{0}".format(key)
         self.multicall_add(m, "d.get_{0}".format(field))
@@ -323,7 +323,7 @@ class Torrent:
         """
 
         self._assert_custom_key_valid(key)
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
 
         self.multicall_add(m, "d.set_custom{0}".format(key), value)
 
@@ -353,7 +353,7 @@ class Torrent:
         """Check if torrent is waiting to be hash checked
 
         @note: Variable where the result for this method is stored Torrent.hash_checking_queued"""
-        m = rtorrent.rpc.Multicall(self)
+        m = rtorrentlib.rpc.Multicall(self)
         self.multicall_add(m, "d.get_hashing")
         self.multicall_add(m, "d.is_hash_checking")
         results = m.call()